bugfix # ??? : fixed referrer issue fot HTML5 library.
Currently the html5 library does its own domain restriction check. It then send requests to Kaltura server which get the iframe location as the referrer.
Because domain referrer checking is not a "solid" security check anyway we can rely on the html5 library check and not check it within the Kaltura server.


git-svn-id: svn+ssh://kelev.kaltura.com/usr/local/kalsource/backend/server/trunk/core@61095 6b8eccd3-e8c5-4e7d-8186-e12b5326b719