serve actions caching bug fix
the path of the file to return may be in the cache folder, since the api
cache is shared, some other machine may try to return from cache while
it does not have the locally cached file. The fix is simply not to
return from cache in case the file to serve is not found.