api cache bug fix
this->_responseMetadata should always be unserialized, need to serialize
it right before saving to memcache. In the previous implementation, after
storeCache it was serialized. In cases where a cache was extended (e.g.
from central cache to local cache) it doesn't go through storeCache and
therefore the unserialized array was saved without serialization.