query cache: when caching use the hydration time
if there is a long api session (e.g. thumbnail stripe) and some object
that it uses is updated during execution, an invalid query may be
cached.
the reason is that the object is hydrated only once, due to propel
instance pool. the api may not cache the query at first, since it's too
close to the update time. after some time passes, the api may cache the
query if the same object is pulled again, but the cached object will be
old (hydrated long time ago).
since the query cache currently saves the current server time with the
query, subsequent calls may use the cached query even though it's
invalid.
the fix is to save the oldest hydration time as the 'query time', if
such a scenario happens, subsequent api calls will know the cached query
is invalid, and won't use it.