simulive mpeg-ts wraparound fix
when there's an mpeg-ts timestamp wraparound, the player fails to sync
the captions to the video, and the captions stop. this change moves the
clipStartTime so that it will be far from the wraparound value (1<<33).
as long as the duration of the simulive video is less than 13.5h, no
wraparound will occur.